+++ packages/blade/src/components/Carousel/CarouselItem.web.tsx
@@ -5,9 +5,11 @@ import styled from 'styled-components';
 import React from 'react';
 import type { CarouselProps } from './types';
 import { useCarouselContext } from './CarouselContext';
+import { componentIds } from './constants';
 import BaseBox from '~components/Box/BaseBox';
 import { useBreakpoint, useTheme } from '~utils';
 import { makeAccessible } from '~utils/makeAccessible';
+import { assignWithoutSideEffects } from '~utils/assignWithoutSideEffects';
 
 type StyledCarouselItemProps = Pick<CarouselProps, 'visibleItems' | 'shouldAddStartEndSpacing'> &
   Pick<CarouselItemProps, 'shouldHaveEndSpacing' | 'shouldHaveStartSpacing'> & {
@@ -48,7 +50,7 @@ type CarouselItemProps = {
   shouldHaveEndSpacing?: boolean;
 };
 
-const CarouselItem = ({
+const _CarouselItem = ({
   children,
   shouldHaveStartSpacing,
   shouldHaveEndSpacing,
@@ -88,5 +90,9 @@ const CarouselItem = ({
     </StyledCarouselItem>
   );
 };
+
+const CarouselItem = assignWithoutSideEffects(_CarouselItem, {
+  componentId: componentIds.CarouselItem,
+});
 
 export { CarouselItem };