+++ packages/blade/src/components/Input/DropdownInputTriggers/__tests__/AutoComplete.ssr.test.tsx
@@ -16,14 +16,6 @@ import { Box } from '~components/Box';
  */
 
 describe('<Dropdown /> with <AutoComplete />', () => {
-  afterAll(() => {
-    // These are not defined by default in JSDOM so clearing them out.
-    // @ts-expect-error: it is taking web's requestAnimationFrame types but JSDom doesn't define these
-    global.requestAnimationFrame = null;
-    // @ts-expect-error: it is expecting web's requestAnimationFrame types but JSDom doesn't define these
-    global.cancelAnimationFrame = null;
-  });
-
   it('should render dropdown and make it visible on click', async () => {
     const { container, getByRole, queryByRole } = renderWithSSR(
       <Dropdown>
@@ -42,12 +34,6 @@ describe('<Dropdown /> with <AutoComplete />', () => {
         </DropdownOverlay>
       </Dropdown>,
     );
-
-    // Cannot define this in beforeEach because we want it to be defined after renderToString call
-    // @ts-expect-error: too lazy to define accurate typescript mocks just for mocking
-    global.requestAnimationFrame = (cb) => cb();
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    global.cancelAnimationFrame = () => {};
 
     const selectInput = getByRole('combobox', { name: 'Fruits' });